+++ vscode-extension/mtlog-analyzer/src/test/fixtures/quickfix_args.go
@@ -1,0 +1,8 @@
+package fixtures
+
+import "github.com/willibrandon/mtlog"
+
+func TestQuickFixArgs() {
+	logger := mtlog.New(mtlog.WithConsole())
+	logger.Information("User {UserId} action {Action}", 123)
+}